Remove ConstantFolding
