+++ src/test/data/None.txt
@@ -37,13 +37,13 @@ rand:4/6,3/6
 input:
 2d6+1-1-2-3-4
 output:
-DiceBot : (2D6-9) ＞ 5[4,1]-9 ＞ -4
+DiceBot : (2D6+1-1-2-3-4) ＞ 5[4,1]+1-1-2-3-4 ＞ -4
 rand:4/6,1/6
 ============================
 input:
 2d10--3-4
 output:
-DiceBot : (2D10-1) ＞ 8[3,5]-1 ＞ 7
+DiceBot : (2D10+3-4) ＞ 8[3,5]+3-4 ＞ 7
 rand:3/10,5/10
 ============================
 input:
@@ -61,13 +61,13 @@ rand:2/6,4/6
 input:
 2d10+3-4
 output:
-DiceBot : (2D10-1) ＞ 8[3,5]-1 ＞ 7
+DiceBot : (2D10+3-4) ＞ 8[3,5]+3-4 ＞ 7
 rand:3/10,5/10
 ============================
 input:
 2d10+3*4
 output:
-DiceBot : (2D10+12) ＞ 8[3,5]+12 ＞ 20
+DiceBot : (2D10+3*4) ＞ 8[3,5]+3*4 ＞ 20
 rand:3/10,5/10
 ============================
 input:
