+++ src/test/data/dummyBot.txt
@@ -96,13 +96,13 @@ rand:
 input:
 2D6*2*3
 output:
-DiceBot : (2D6*6) ＞ 5[2,3]*6 ＞ 30
+DiceBot : (2D6*2*3) ＞ 5[2,3]*2*3 ＞ 30
 rand:2/6,3/6
 ============================
 input:
 2D6*2*3-1
 output:
-DiceBot : (2D6*6-1) ＞ 5[2,3]*6-1 ＞ 29
+DiceBot : (2D6*2*3-1) ＞ 5[2,3]*2*3-1 ＞ 29
 rand:2/6,3/6
 ============================
 input:
